Add post-operative frontend
