+++ src/main/java/com/project/vetdata/controller/UserController.java
@@ -3,7 +3,6 @@ package com.project.vetdata.controller;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 
 @Controller
